调试中途的版本，现在按下回原点的按键会真的回到原点，原因未知

+++ stm32f103c8t6/Core/Inc/Coordinate.h
@@ -19,8 +19,8 @@ typedef enum
 //二维像素点uint16_t
 typedef struct 
 {
-    uint16_t x;
-    uint16_t y;
+    int32_t x;
+    int32_t y;
 } Pixel_Point;
 
 //二维毫米点float
